refactor(animation): upgrade `@halvaradop/tailwindcss-animations`
