+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "dev": "turbo dev --parallel --filter=!*page",
     "dev:web": "turbo dev --filter=*page",
-    "build": "pnpm build:core && turbo run build --parallel --filter=!*core",
+    "build": "pnpm build:core && turbo run build --parallel --filter=!*core --filter=!*merge",
     "build:web": "turbo run build --filter=*page",
     "build:core": "turbo run build --filter=*core",
     "format": "prettier --write .",
@@ -45,6 +45,7 @@
   "devDependencies": {
     "@csstools/postcss-minify": "^2.0.3",
     "@halvaradop/ts-utility-types": "^0.19.0",
+    "@tailwindcss/cli": "^4.1.4",
     "@types/node": "^22.14.0",
     "@vitest/coverage-v8": "^3.1.1",
     "autoprefixer": "^10.4.21",
