+++ packages/tailwindcss-animations/src/theme.js
@@ -1,0 +1,59 @@
+export const theme = {
+    animationIterationCount: {
+        none: "0",
+        once: "1",
+        twice: "2",
+        thrice: "3",
+        infinite: "infinite",
+    },
+    animationFillMode: {
+        none: "none",
+        forwards: "forwards",
+        backwards: "backwards",
+        both: "both",
+    },
+    animationRange: {
+        normal: "normal",
+        cover: "cover",
+        contain: "contain",
+        entry: "entry",
+        exit: "exit",
+        gradual: "10% 90%",
+        moderate: "20% 80%",
+        brisk: "30% 70%",
+        rapid: "40% 60%",
+    },
+    timeline: {
+        none: "none",
+        auto: "auto",
+        single: "--single-timeline",
+        scroll: "scroll()",
+        view: "view()",
+    },
+    scrollTimeline: {
+        single: "--single-timeline",
+    },
+    viewTimeline: {
+        single: "--single-timeline",
+    },
+    scrollTimelineAxis: {
+        block: "block",
+        inline: "inline",
+        x: "x",
+        y: "y",
+    },
+    viewTimelineAxis: {
+        block: "block",
+        inline: "inline",
+        x: "x",
+        y: "y",
+    },
+    scrollAnimation: {
+        single: "--single-timeline",
+    },
+    viewAnimation: {
+        single: "--single-timeline",
+    },
+}
+
+export default theme
