--- packages/tailwindcss-animations/src/utils.ts
@@ -1,11 +1,0 @@
-/**
- * Match utilities that are generated by matchUtilities function provided by TailwindCSS.
- * @example
- *
- * // Expected: true
- * matchUtiltiesRegex.test("animation-delay-100")
- *
- * // Expected: false
- * matchUtiltiesRegex.test("animation")
- */
-export const matchUtilitiesRegex = new RegExp(/^(animation(?=Delay|Duration|Steps|IterationCount|FillMode|Range))/)
