--- packages/tailwindcss-utilities/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@halvaradop/tailwindcss-core/tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["dist"]
-}
